requirements.txt
mysql-connector==2.1.6
beautifulsoup4==4.7.1
